# Optimize Circuit Reduction and Replace 2-Port Ground Network with 1-Port Short Network