--- a/skrf/circuit.py
+++ b/skrf/circuit.py
@@ -743,11 +743,13 @@ class Circuit:
         There is a numerical bottleneck in this function,
         when creating the block diagonal matrice [X] from the [X]_k matrices.
         """
-        idx, Xf = 0, np.zeros((len(self.frequency), self.dim, self.dim), dtype='complex')
-        for cnx in self.connections:
-            idx_s, idx_e = idx, idx + len(cnx)
-            Xf[:, idx_s:idx_e, idx_s:idx_e] = self._Xk(cnx)
-            idx = idx_e
+        Xks = [self._Xk(cnx) for cnx in self.connections]
+
+        Xf = np.zeros((len(self.frequency), self.dim, self.dim), dtype='complex')
+        off = np.array([0, 0])
+        for Xk in Xks:
+            Xf[:, off[0]:off[0] + Xk.shape[1], off[1]:off[1]+Xk.shape[2]] = Xk
+            off += Xk.shape[1:]
 
         return Xf
 
@@ -1509,6 +1511,10 @@ def reduce_circuit(connections: list[list[tuple[Network, int]]],
                 tmp_gnd = Circuit.Ground(frequency=ground_ntwk.frequency,
                                          name=f'G_{ntwk.name}_{port}',
                                          z0=ground_ntwk.z0)
+
+                # Transfet the ground Network to a short Network
+                tmp_gnd = tmp_gnd.subnetwork(ports=(0,))
+
                 tmp_cnxs.append([(ntwk, port), (tmp_gnd, 0)])
 
         connections = tmp_cnxs