## Introduction
The concept of `Circuit.Ground` was introduced in https://github.com/scikit-rf/scikit-rf/issues/626:
> A `Circuit.Ground` is indeed a 2-port Network, as it models an infinite shunt admittance.

Using `Ground` in the `Circuit` seems straightforward. However, the issue also mentions that if you only want to short one port, it's better to **short it and connect it directly**.

## Optimization through Short Networks
1. Considering that `short Network` is a 1-port Network, connecting a short with an `N-Port Network` results in an `(N-1)-Port network`. This means that using a `short Network` will gradually reduce the total number of ports in the `Circuit` involved in the calculation while obtaining the same result.
2. In the `reduce_circuit()` method, we observed that performance can be improved by handling Ground specifically, as shown in https://github.com/scikit-rf/scikit-rf/pull/1080. In which, we added a mechanism to **automatically handle `Ground Network`**.

Therefore, I thought it would be an interesting idea to automatically convert `Ground` to `short` in the `reduce_circuit()` method.

## Benchmark Results
Benchmark code can be found in https://github.com/scikit-rf/scikit-rf/pull/1080, and the results are as follows:
```bash

(skrf) (base) ➜  scikit-rf git:(master) ✗ python circuit_performance.py
skrf path: /home/lansus/WorkFile/Git/scikit-rf/skrf/__init__.py
Total time in Easy example: 3.8372 ms, min: 3.6426 ms, std: 0.2304 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 8.4084 ms, min: 7.9572 ms, std: 0.3509 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 46.3497 ms, min: 43.3400 ms, std: 2.7975 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 41.2589 ms, min: 38.5571 ms, std: 1.7028 ms
Standard (47) and Optimized (13) results are equal: True

(skrf) (base) ➜  scikit-rf git:(master) ✗ git switch CircuitPerformance
(skrf) (base) ➜  scikit-rf git:(CircuitPerformance) ✗ python circuit_performance.py   
skrf path: /home/lansus/WorkFile/Git/scikit-rf/skrf/__init__.py
Total time in Easy example: 4.1243 ms, min: 3.7518 ms, std: 0.4754 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 7.5973 ms, min: 7.0727 ms, std: 0.4545 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 44.6984 ms, min: 41.3613 ms, std: 2.6458 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 24.1429 ms, min: 23.2019 ms, std: 0.7935 ms
Standard (47) and Optimized (13) results are equal: True
```
## Conclusion
Automatically converting `Ground` to `short` in the `reduce_circuit()` method can effectively reduce the total number of ports involved in calculations, resulting in performance improvements. The benchmark results demonstrate that this approach maintains accuracy while optimizing computation time.