@Asachoo 2024-07-30
Using `media.short` directly in the Circuit doesn't seem to cause any issues.
Therefore, I don't quite understand why a 2-Port Network `Circuit.Ground`, which models an infinite shunt admittance, is used to represent ground port in the circuit. Could you please explain the benefits of using a `2-port Ground Network`? @jhillairet , your insights would be greatly appreciated.