@Asachoo 2024-07-30
Thank you for providing the information from the article @jhillairet.
Based on this, I believe the internal implementation of `Ground`/`Open` in `Circuit` can actually be replaced with `media.open`/`media.short`. This would result in a more consistent definition, a more reasonable Number of Network's Ports, and better circuit performance.
I plan to implement the changes. Any thoughts or feedback from you or anyone else would be greatly appreciated.